Add "Legacy" story sample code
By force of habit I'll keep the commit messages in English, but all the content
within the repository will be intentionally written in Spanish.

This changest includes all the code related to the "Legacy" story, including:

- The legacy mock app
- The Operator logic using kopf framework
- The Kubernetes manifest used for provisioning everything

I'll invest a litte bit of more time elaborating on the docs and it will go
on a separate change set.

Signed-off-by: Jonnatan Jossemar Cordero <jossemargt@gmail.com>